requirements.txt
flask==1.1.2
markdown2==2.3.10
pyyaml==5.3.1
beautifulsoup4==4.9.3
Unidecode==1.1.1
matplotlib==3.3.3
scipy==1.5.4
colour==0.1.5
